+++ src/App.js
@@ -1,13 +1,13 @@
 /* eslint-disable react/prefer-stateless-function */
 import React from 'react';
 // import Calculator from './components/calculator';
-import Calculator2 from './components/calculator2';
+import Calculator from './components/calculator';
 import './styles.css';
 
 class App extends React.Component {
   render() {
     return (
-      <Calculator2 />
+      <Calculator />
     );
   }
 }
