+++ src/components/calculator.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React, { Fragment, useState } from 'react';
 import calculate from '../logic/calculate';
 
 const buttonsRows = [
@@ -9,43 +9,43 @@ const buttonsRows = [
   ['0', '.', '='],
 ];
 
-export default class Calculator extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      total: null,
-      next: null,
-      operation: null,
-    };
-  }
+const defaultCalculatorData = {
+  total: null,
+  next: null,
+  operation: null,
+};
 
-  handleCalculation = (buttonName) => {
-    const updatedState = calculate(this.state, buttonName);
-    this.setState(updatedState);
-  }
+const Calculator = () => {
+  const [calculatorData, setCalculatorData] = useState(defaultCalculatorData);
+  const { total, next, operation } = calculatorData;
 
-  render() {
-    const { total, next, operation } = this.state;
-    const isStateNull = () => !total && !next && !operation;
-    const displayCalculation = () => {
-      if (operation) return `${total} ${operation} ${next || ''}`;
-      return next || total;
-    };
-    return (
-      <div className="calculator-grid">
-        <div className="output">{isStateNull() ? '0' : displayCalculation()}</div>
-        {
+  const handleCalculation = (buttonName) => {
+    const { total, next, operation } = calculate(calculatorData, buttonName);
+    setCalculatorData({ total, next, operation });
+  };
+
+  const isStateNull = () => !total && !next && !operation;
+
+  const displayCalculation = () => {
+    if (operation) return `${total} ${operation} ${next || ''}`;
+    return next || total;
+  };
+
+  return (
+    <div className="calculator-grid">
+      <div className="output">{isStateNull() ? '0' : displayCalculation()}</div>
+      {
           buttonsRows.map((buttonsRow, rowIndex) => (
             <Fragment key={`row ${rowIndex + 1}`}>
               {
                   buttonsRow.map((buttonName, index) => (
                     <button
-                      onClick={() => { this.handleCalculation(buttonName); }}
+                      onClick={() => { handleCalculation(buttonName); }}
                       type="button"
                       key={buttonName}
                       className={`calc-btn 
-                      ${buttonName === '0' ? 'span-two' : ''} 
-                      ${index === buttonsRow.length - 1 ? 'operator' : ''}`}
+                                ${buttonName === '0' ? 'span-two' : ''} 
+                                ${index === buttonsRow.length - 1 ? 'operator' : ''}`}
                     >
                       { buttonName }
                     </button>
@@ -54,8 +54,8 @@ export default class Calculator extends React.Component {
             </Fragment>
           ))
         }
-      </div>
+    </div>
+  );
+};
 
-    );
-  }
-}
+export default Calculator;
